// 公司/data100/kbt-console/src/pages/adv/ProjectManager/AddTask/Subtasks/data.d.ts
export interface TableListItem {
  address: string;
  city: string;
  finishStat: string;
  taskId: string;
  phone: string;
  // brand: string;
  // pointFile: string;
  // pointUploadMsg: string;
  // pointProgress: number;
  // putStartDate: string;
  // putEndDate: string;
  // execStartDate: string;
  // execEndDate: string;
  // taskStatus: string;
  // auditProgress: string;
}

export interface QueryParams {
  address: string;
  city: string;
  finishStat: string;
  taskId: string;
  phone: string;

}